qapi: store raw expressions to QAPISchema
After cleaning up the comments of qapi-schema.json file, we get
an range of raw expressions, they are also converted to ordered
dictionaries.

This patch just addes a member to QAPISchema to store the raw
expressions.

Actually this patch was split from QMP introspection patchset,
it's no longer needed in that patchset. I would like to post
this patch singly, we can apply it if it's useful.

Signed-off-by: Amos Kong <akong@redhat.com>