qapi: introduce DataObject to describe dynamic structs
This patch introduced a DataObject union in qapi-schema.json,
we use it to describe dynamic data structs.

We will use it in following patches to support to QMP full
introspection. We have many kinds of schema in json file,
they all can be described by DataObject.

This patch also added a doc: qmp-full-introspection.txt,
QMP introspection releated document will be added into it.
It helps to use the new query command and understand the
abstract method in describing the dynamic struct.

Signed-off-by: Amos Kong <akong@redhat.com>